+++ lib/module/dialog-flow.component.d.ts
@@ -1,7 +1,7 @@
 import { DialogFlowFulfillmentResponse } from '../interfaces/dialog-flow-fulfillment-response.interface';
 import { DialogFlowResponse } from '../interfaces/dialog-flow-response.interface';
 import { HandlerContainer } from './../core';
-export declare class DialogFlowService {
+export declare class DialogFlowComponent {
     private readonly handlerContainer;
     constructor(handlerContainer: HandlerContainer);
     handleIntentOrAction(dialogFlowResponse: DialogFlowResponse): Promise<DialogFlowFulfillmentResponse>;