+++ samples/01-dialogflow-handlers/src/app.component2.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class AppComponent2 {
+    public haveBeenCalled(intentOrAction: string): void {
+        console.log(`intent or action ${intentOrAction} have been called`);
+    }
+}