+++ src/module/dialog-flow.component.ts
@@ -1,10 +1,10 @@
-import { Injectable, Provider } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { DialogFlowFulfillmentResponse } from '../interfaces/dialog-flow-fulfillment-response.interface';
 import { DialogFlowResponse } from '../interfaces/dialog-flow-response.interface';
 import { HandlerContainer } from './../core';
 
 @Injectable()
-export class DialogFlowService {
+export class DialogFlowComponent {
 	constructor(private readonly handlerContainer: HandlerContainer) {}
 
 	public async handleIntentOrAction(
@@ -13,7 +13,10 @@ export class DialogFlowService {
 		const intent = dialogFlowResponse.queryResult.intent.displayName;
 		const action = dialogFlowResponse.queryResult.action;
 
-		const fulfillment = this.handlerContainer.findAndCallHandlers(intent, dialogFlowResponse);
+		const fulfillment = this.handlerContainer.findAndCallHandlers(dialogFlowResponse, {
+			intent,
+			action,
+		});
 		return fulfillment as DialogFlowFulfillmentResponse;
 	}
 }