+++ src/module/dialog-flow.controller.ts
@@ -1,12 +1,12 @@
 import { Body, Controller, HttpStatus, RequestMethod, Res } from '@nestjs/common';
 import { DialogFlowResponse } from '../interfaces/dialog-flow-response.interface';
-import { DialogFlowService } from './dialog-flow.component';
+import { DialogFlowComponent } from './dialog-flow.component';
 import { METHOD_METADATA, PATH_METADATA } from '../constant';
 import { WebHookConfig } from '../interfaces/web-hook-config.interface';
 
 @Controller()
 export class DialogFlowController {
-	constructor(private readonly dialogFlowService: DialogFlowService) {}
+	constructor(private readonly DialogFlowComponent: DialogFlowComponent) {}
 
 	public static forRoot(webHookConfig: WebHookConfig) {
 		Reflect.defineMetadata(PATH_METADATA, webHookConfig.basePath, DialogFlowController);
@@ -24,7 +24,7 @@ export class DialogFlowController {
 	}
 
 	async dialogFlowWebHook(@Body() dialogFlowResponse: DialogFlowResponse, @Res() res) {
-		const fulfillment = await this.dialogFlowService.handleIntentOrAction(dialogFlowResponse);
+		const fulfillment = await this.DialogFlowComponent.handleIntentOrAction(dialogFlowResponse);
 		return res.status(HttpStatus.OK).send(fulfillment);
 	}
 }